msm: scm: Add atomic SCM APIs
The atomic SCM APIs are useful for commands that are guaranteed
by the secure side to be uninterruptable, atomic and SMP safe.
The calling convention use registers for passing parameters and
return values between the secure and non-secure side. Support
this interface with  scm_call_atomic[1-2]() functions
corresponding to the number of arguments passed.

Change-Id: I4746d6c821522c49f6aaa7234cdd99f9378f3006
Signed-off-by: Stephen Boyd <sboyd@codeaurora.org>